- add specialization design in headerSection - delete home from navabr
